+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.0 (in development)
 * Add the `canu.show_history_page` method.
+* Add the `canu.EventHandler.redundant` attribute.
 
 ## 0.6.0 (2024-06-09)
 * Add the `canu.EventHandler.on_event` method.