+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.0 (in development)
 * Add the `canu.show_history_page` method.
 * Add the `canu.EventHandler.redundant` attribute.
+* Add the `canu.handle_files` method.
 
 ## 0.6.0 (2024-06-09)
 * Add the `canu.EventHandler.on_event` method.